pyproject.toml
[tool.poetry]
name = "lineageos_updates_bot"
version = "1.0.0"
description = "LineageOS updates Telegram bot"
authors = ["Sebastiano Barezzi <seba@sebaubuntu.dev>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/sebaubuntu-python/lineageos_updates_bot"

[tool.poetry.dependencies]
python = "^3.8"
humanize = "^4.1.0"
liblineage = "^1.1.0"
python-telegram-bot = "^20.0a4"
sebaubuntu-libs = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
